% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NOAASigEarthqkDb_functions.R
\name{eq_clean_data}
\alias{eq_clean_data}
\title{eq_clean_data}
\usage{
eq_clean_data(filename)
}
\arguments{
\item{filename}{The name of the file that the function will read into R.}
}
\value{
This function removes events in the BCE era (49 out of the 6416 events) and returns a cleaned NOAA data
frame, with the following additional changes:
\enumerate{
\item A Date column created by uniting the year, month, and day variables, and converting it to the Date class
\item Latitude and Longitude variables converted to numeric class
}
}
\description{
This function further cleans the NOAA data frame produced by the eq_location_clean function.
}
\note{
Example: eq_clean_data("./inst/extdata/earthquakes_-2150to2024_clean1.csv")

If the file  does not exist (e.g. earthquakes_-2150to2050.csv), attempting to read this file
will generate an error.
try(eq_clean_data("./inst/extdata/earthquakes_-2150to2050_clean1.csv"))

Attempting to read a non-CSV file will also generate an error.
try(eq_clean_data("./inst/extdata/earthquakes_-2150to2024_clean1.xlsx"))
}
